requirements.txt
appdirs==1.4.4
async-timeout==4.0.3
build==1.2.1
buildozer==1.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
Cython==0.29.33
distlib==0.3.8
docutils==0.21.2
filelock==3.15.4
idna==3.7
ifaddr==0.2.0
Jinja2==3.1.4
Kivy==2.3.0
Kivy-Garden==0.1.5
MarkupSafe==2.1.5
packaging==24.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
ptyprocess==0.7.0
Pygments==2.18.0
pyproject_hooks==1.1.0
pyzmq==26.0.3
requests==2.32.3
sh==1.14.3
toml==0.10.2
tomli==2.0.1
urllib3==2.2.2
virtualenv==20.26.3
zeroconf==0.132.2
